requirements.txt
flask
spacy
nltk
PyPDF2
python-docx
